Move BackingStore into store module
