+++ src/merk/mod.rs
@@ -1,4 +1,3 @@
-pub mod backingstore;
 mod client;
 #[cfg(feature = "merk-full")]
 mod ics23;
@@ -7,10 +6,10 @@ mod proofbuilder;
 #[cfg(feature = "merk-full")]
 pub mod store;
 
-pub use backingstore::{BackingStore, ProofStore};
 pub use client::Client;
 pub use merk;
 #[cfg(feature = "merk-full")]
 pub use proofbuilder::ProofBuilder;
 #[cfg(feature = "merk-full")]
 pub use store::MerkStore;
+pub use store::ProofStore;
