+++ src/merk/store.rs
@@ -2,12 +2,14 @@ use crate::abci::ABCIStore;
 use crate::error::{Error, Result};
 use crate::store::*;
 use merk::{
-    chunks::ChunkProducer, restore::Restorer, rocksdb, tree::Tree, BatchEntry, Hash, Merk, Op,
+    chunks::ChunkProducer, proofs::query::Map as ProofMap, restore::Restorer, rocksdb, tree::Tree,
+    BatchEntry, Hash, Merk, Op,
 };
 use std::cell::RefCell;
 use std::{collections::BTreeMap, convert::TryInto};
 use std::{
     mem::transmute,
+    ops::Bound,
     path::{Path, PathBuf},
 };
 use tendermint_proto::v0_34::abci::{self, *};
@@ -485,4 +487,40 @@ fn read_u64(bytes: &[u8]) -> u64 {
     let mut array = [0; 8];
     array.copy_from_slice(bytes);
     u64::from_be_bytes(array)
+}
+
+pub struct ProofStore(pub ProofMap);
+
+impl Read for ProofStore {
+    fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
+        let maybe_value = self.0.get(key).map_err(|err| {
+            if let merk::Error::MissingData = err {
+                Error::StoreErr(crate::store::Error::ReadUnknown(key.to_vec()))
+            } else {
+                Error::Merk(err)
+            }
+        })?;
+        Ok(maybe_value.map(|value| value.to_vec()))
+    }
+
+    fn get_next(&self, key: &[u8]) -> Result<Option<KV>> {
+        let mut iter = self.0.range((Bound::Excluded(key), Bound::Unbounded));
+        let item = iter.next().transpose().map_err(|err| {
+            if let merk::Error::MissingData = err {
+                Error::StoreErr(crate::store::Error::ReadUnknown(key.to_vec()))
+            } else {
+                Error::Merk(err)
+            }
+        })?;
+        Ok(item.map(|(k, v)| (k.to_vec(), v.to_vec())))
+    }
+
+    fn get_prev(&self, key: Option<&[u8]>) -> Result<Option<KV>> {
+        let mut iter = self.0.range((
+            Bound::Unbounded,
+            key.map_or(Bound::Unbounded, Bound::Excluded),
+        ));
+        let item = iter.next_back().transpose()?;
+        Ok(item.map(|(k, v)| (k.to_vec(), v.to_vec())))
+    }
 }
