+++ src/migrate.rs
@@ -130,7 +130,7 @@ mod tests {
         encoding::{Decode, Encode},
         orga,
         state::State,
-        store::{DefaultBackingStore, MapStore, Read, Shared, Store, Write},
+        store::{BackingStore, DefaultBackingStore, MapStore, Read, Shared, Store, Write},
         Result,
     };
 
@@ -212,10 +212,7 @@ mod tests {
     }
 
     fn create_foo_v0_store() -> Result<Store> {
-        #[cfg(any(feature = "merk", feature = "merk-verify"))]
-        let mut store = Store::new(DefaultBackingStore::MapStore(Shared::new(MapStore::new())));
-        #[cfg(all(not(feature = "merk"), not(feature = "merk-verify")))]
-        let mut store = Store::new(DefaultBackingStore::new(MapStore::new()));
+        let mut store = Store::new(BackingStore::MapStore(Shared::new(MapStore::new())));
 
         let mut foo = FooV0 {
             bar: 42,
