+++ src/store/backingstore.rs
@@ -1,9 +1,10 @@
+#[cfg(feature = "merk")]
+use crate::merk::ProofStore;
 #[cfg(feature = "merk-full")]
-use super::{MerkStore, ProofBuilder};
+use crate::merk::{MerkStore, ProofBuilder};
+use crate::store::ReadWrite;
 use crate::store::{bufstore::PartialMapStore, BufStore, Empty, MapStore, Read, Shared, Write, KV};
 use crate::{Error, Result};
-use merk::proofs::query::Map as ProofMap;
-use std::any::Any;
 use std::ops::Bound;
 
 #[cfg(feature = "merk-full")]
@@ -11,27 +12,19 @@ type WrappedMerkStore = Shared<BufStore<Shared<BufStore<Shared<MerkStore>>>>>;
 
 #[derive(Clone)]
 pub enum BackingStore {
+    MapStore(Shared<MapStore>),
+    PartialMapStore(Shared<PartialMapStore>),
+    Null(Empty),
+    Other(Shared<Box<dyn ReadWrite>>),
+
     #[cfg(feature = "merk-full")]
     WrappedMerk(WrappedMerkStore),
     #[cfg(feature = "merk-full")]
     ProofBuilder(ProofBuilder),
     #[cfg(feature = "merk-full")]
     Merk(Shared<MerkStore>),
-    MapStore(Shared<MapStore>),
-    PartialMapStore(Shared<PartialMapStore>),
+    #[cfg(feature = "merk")]
     ProofMap(Shared<ProofStore>),
-    Null(Empty),
-    Other(Shared<Box<dyn ReadWrite>>),
-}
-
-pub trait ReadWrite: Read + Write + Any + 'static {
-    fn into_any(self: Box<Self>) -> Box<dyn Any>;
-}
-
-impl<T: Read + Write + 'static> ReadWrite for T {
-    fn into_any(self: Box<Self>) -> Box<dyn Any> {
-        self
-    }
 }
 
 impl Default for BackingStore {
@@ -43,49 +36,55 @@ impl Default for BackingStore {
 impl Read for BackingStore {
     fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
         match self {
+            BackingStore::MapStore(ref store) => store.get(key),
+            BackingStore::PartialMapStore(ref store) => store.get(key),
+            BackingStore::Null(ref null) => null.get(key),
+            BackingStore::Other(ref store) => store.borrow().get(key),
+
             #[cfg(feature = "merk-full")]
             BackingStore::WrappedMerk(ref store) => store.get(key),
             #[cfg(feature = "merk-full")]
             BackingStore::ProofBuilder(ref builder) => builder.get(key),
             #[cfg(feature = "merk-full")]
             BackingStore::Merk(ref store) => store.get(key),
-            BackingStore::MapStore(ref store) => store.get(key),
-            BackingStore::PartialMapStore(ref store) => store.get(key),
+            #[cfg(feature = "merk")]
             BackingStore::ProofMap(ref map) => map.get(key),
-            BackingStore::Null(ref null) => null.get(key),
-            BackingStore::Other(ref store) => store.borrow().get(key),
         }
     }
 
     fn get_next(&self, key: &[u8]) -> Result<Option<KV>> {
         match self {
+            BackingStore::MapStore(ref store) => store.get_next(key),
+            BackingStore::PartialMapStore(ref store) => store.get_next(key),
+            BackingStore::Null(ref null) => null.get_next(key),
+            BackingStore::Other(ref store) => store.borrow().get_next(key),
+
             #[cfg(feature = "merk-full")]
             BackingStore::WrappedMerk(ref store) => store.get_next(key),
             #[cfg(feature = "merk-full")]
             BackingStore::ProofBuilder(ref builder) => builder.get_next(key),
             #[cfg(feature = "merk-full")]
             BackingStore::Merk(ref store) => store.get_next(key),
-            BackingStore::MapStore(ref store) => store.get_next(key),
-            BackingStore::PartialMapStore(ref store) => store.get_next(key),
+            #[cfg(feature = "merk")]
             BackingStore::ProofMap(ref map) => map.get_next(key),
-            BackingStore::Null(ref null) => null.get_next(key),
-            BackingStore::Other(ref store) => store.borrow().get_next(key),
         }
     }
 
     fn get_prev(&self, key: Option<&[u8]>) -> Result<Option<KV>> {
         match self {
+            BackingStore::MapStore(ref store) => store.get_prev(key),
+            BackingStore::PartialMapStore(ref store) => store.get_prev(key),
+            BackingStore::Null(ref null) => null.get_prev(key),
+            BackingStore::Other(ref store) => store.borrow().get_prev(key),
+
             #[cfg(feature = "merk-full")]
             BackingStore::WrappedMerk(ref store) => store.get_prev(key),
             #[cfg(feature = "merk-full")]
             BackingStore::ProofBuilder(ref builder) => builder.get_prev(key),
             #[cfg(feature = "merk-full")]
             BackingStore::Merk(ref store) => store.get_prev(key),
-            BackingStore::MapStore(ref store) => store.get_prev(key),
-            BackingStore::PartialMapStore(ref store) => store.get_prev(key),
+            #[cfg(feature = "merk")]
             BackingStore::ProofMap(ref map) => map.get_prev(key),
-            BackingStore::Null(ref null) => null.get_prev(key),
-            BackingStore::Other(ref store) => store.borrow().get_prev(key),
         }
     }
 }
@@ -93,6 +92,11 @@ impl Read for BackingStore {
 impl Write for BackingStore {
     fn put(&mut self, key: Vec<u8>, value: Vec<u8>) -> Result<()> {
         match self {
+            BackingStore::MapStore(ref mut store) => store.put(key, value),
+            BackingStore::PartialMapStore(ref mut store) => store.put(key, value),
+            BackingStore::Null(ref mut store) => store.put(key, value),
+            BackingStore::Other(ref mut store) => store.borrow_mut().put(key, value),
+
             #[cfg(feature = "merk-full")]
             BackingStore::WrappedMerk(ref mut store) => store.put(key, value),
             #[cfg(feature = "merk-full")]
@@ -101,33 +105,31 @@ impl Write for BackingStore {
             BackingStore::ProofBuilder(_) => {
                 panic!("put() is not implemented for ProofBuilder")
             }
-            BackingStore::MapStore(ref mut store) => store.put(key, value),
-            BackingStore::PartialMapStore(ref mut store) => store.put(key, value),
+            #[cfg(feature = "merk")]
             BackingStore::ProofMap(_) => {
                 panic!("put() is not implemented for ProofMap")
             }
-            BackingStore::Null(ref mut store) => store.put(key, value),
-            BackingStore::Other(ref mut store) => store.borrow_mut().put(key, value),
         }
     }
     fn delete(&mut self, key: &[u8]) -> Result<()> {
         match self {
+            BackingStore::MapStore(ref mut store) => store.delete(key),
+            BackingStore::PartialMapStore(ref mut store) => store.delete(key),
+            BackingStore::Null(ref mut store) => store.delete(key),
+            BackingStore::Other(ref mut store) => store.borrow_mut().delete(key),
+
             #[cfg(feature = "merk-full")]
             BackingStore::WrappedMerk(ref mut store) => store.delete(key),
             #[cfg(feature = "merk-full")]
             BackingStore::Merk(ref mut store) => store.delete(key),
             #[cfg(feature = "merk-full")]
-            #[cfg(feature = "merk-full")]
             BackingStore::ProofBuilder(_) => {
                 panic!("delete() is not implemented for ProofBuilder")
             }
-            BackingStore::MapStore(ref mut store) => store.delete(key),
-            BackingStore::PartialMapStore(ref mut store) => store.delete(key),
+            #[cfg(feature = "merk")]
             BackingStore::ProofMap(_) => {
                 panic!("delete() is not implemented for ProofMap")
             }
-            BackingStore::Null(ref mut store) => store.delete(key),
-            BackingStore::Other(ref mut store) => store.borrow_mut().delete(key),
         }
     }
 }
@@ -205,41 +207,5 @@ impl From<Shared<MerkStore>> for BackingStore {
 impl From<Shared<MapStore>> for BackingStore {
     fn from(store: Shared<MapStore>) -> BackingStore {
         BackingStore::MapStore(store)
-    }
-}
-
-pub struct ProofStore(pub ProofMap);
-
-impl Read for ProofStore {
-    fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
-        let maybe_value = self.0.get(key).map_err(|err| {
-            if let merk::Error::MissingData = err {
-                Error::StoreErr(crate::store::Error::ReadUnknown(key.to_vec()))
-            } else {
-                Error::Merk(err)
-            }
-        })?;
-        Ok(maybe_value.map(|value| value.to_vec()))
-    }
-
-    fn get_next(&self, key: &[u8]) -> Result<Option<KV>> {
-        let mut iter = self.0.range((Bound::Excluded(key), Bound::Unbounded));
-        let item = iter.next().transpose().map_err(|err| {
-            if let merk::Error::MissingData = err {
-                Error::StoreErr(crate::store::Error::ReadUnknown(key.to_vec()))
-            } else {
-                Error::Merk(err)
-            }
-        })?;
-        Ok(item.map(|(k, v)| (k.to_vec(), v.to_vec())))
-    }
-
-    fn get_prev(&self, key: Option<&[u8]>) -> Result<Option<KV>> {
-        let mut iter = self.0.range((
-            Bound::Unbounded,
-            key.map_or(Bound::Unbounded, Bound::Excluded),
-        ));
-        let item = iter.next_back().transpose()?;
-        Ok(item.map(|(k, v)| (k.to_vec(), v.to_vec())))
     }
 }
